Make the highlighting more noticeable
And uniform: newly added or removed stuff is reversed, whether it's the
full-line change or not. It already happened a few times that I wasn't
noticing the change because it wasn't highlighted clear enough.

Actually I'd like to add support for colorschemes for 256-color
terminals, so that it would look nicer.